LayoutTests/imported/w3c/web-platform-tests/referrer-policy/generic/inheritance/iframe-inheritance-history-about-srcdoc-expected.txt


FAIL History navigation reuses original policy. assert_equals: History navigation reuses original policy. expected (undefined) undefined but got (string) "http://localhost:8800/custom"
PASS New srcdoc iframe uses new policy.
